1) Now dynamically getting data from JSONPlaceholder APIs. 2) Also cleaned up and optimized the code. 3) Made the searchbar sticky!
