--- src/App.js
@@ -1,36 +1,0 @@
-import React, { Component } from 'react';
-import CardList from './CardList';
-import SearchBox from './SearchBox';
-import { robots } from './robots';
-
-//STAE -> the description of your app.
-// an object 
-
-class App extends Component {
-  constructor() {
-    super();
-    this.state = {
-      robots: robots,
-      searchField: ''
-    }
-  }
-
-  onSearchChange = (event) => {
-    this.setState({ searchField: event.target.value });
-  }
-  render () {
-    const filteredRobot = this.state.robots.filter(robot => {
-      return robot.name.toLowerCase().includes(this.state.searchField.toLowerCase());
-    });
-    return (
-      <div className='tc'>
-        <h1>Robot Friends</h1>
-        <SearchBox searchChange={this.onSearchChange} />
-        <CardList robots={filteredRobot}/>
-      </div>
-    );
-  }
-}
-
-
-export default App;
