+++ src/index.js
@@ -1,11 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-import App from './App';
+import App from './containers/App';
 import * as serviceWorker from './serviceWorker';
 import 'tachyons';
-import './App.css'
-//import { robots } from './robots';
 
 
 ReactDOM.render(<App />, document.getElementById('root'));
